feat(schemas): move config schemas to root dir
